Split trimmed curve basis handling
