+++ src/ifc/operations/curve.ts
@@ -11,34 +11,24 @@ import type {
   IfcSegmentIndexSelect,
   IfcTrimmedCurve,
 } from "../generated/schema.ts";
-import {
-  normalizePlacement2D,
-  normalizePlacement3D,
-  type NormalizedPlacement2D,
-  type NormalizedPlacement3D,
-  type NormalizedVec2,
-  type NormalizedVec3,
-} from "../normalize.ts";
 import { ifcToBabylonVector } from "../../engine/ifc-coordinates.ts";
 import type { Vec3 } from "../../types.ts";
+import { resolveConicCurveSegment } from "./curve-conic.ts";
+import { resolveTrimmedCurveSegments } from "./curve-trimmed.ts";
+import type {
+  IndexedPolyCurveResolvedSegment,
+  IndexedPolyCurveSegmentKind,
+  ResolvedCurveSegment,
+} from "./curve-types.ts";
+export type {
+  CurveSegmentKind,
+  IndexedPolyCurveResolvedSegment,
+  IndexedPolyCurveSegmentKind,
+  ResolvedCurveSegment,
+} from "./curve-types.ts";
 
 const DEFAULT_ARC_SEGMENTS = 32;
-const DEFAULT_CONIC_SEGMENTS = 64;
 const EPSILON = 1e-9;
-const FULL_CIRCLE_EPSILON = 1e-6;
-
-export type CurveSegmentKind = "line" | "arc" | "curve";
-export type IndexedPolyCurveSegmentKind = "line" | "arc";
-
-export interface ResolvedCurveSegment {
-  kind: CurveSegmentKind;
-  points: Vec3[];
-}
-
-export interface IndexedPolyCurveResolvedSegment extends ResolvedCurveSegment {
-  kind: IndexedPolyCurveSegmentKind;
-  indices: number[];
-}
 
 type RenderableCurve =
   | IfcPolyline
@@ -87,13 +77,6 @@ function normalize(v: Vec3): Vec3 {
   const len = length(v);
   if (len < EPSILON) return { x: 0, y: 0, z: 0 };
   return scale(v, 1 / len);
-}
-
-function normalizeAngle(angle: number): number {
-  const turn = Math.PI * 2;
-  let result = angle % turn;
-  if (result < 0) result += turn;
-  return result;
 }
 
 function cartesianPointToVec3(point: IfcCartesianPoint): Vec3 {
@@ -170,218 +153,6 @@ function indexedPoints(points: Vec3[], indices: readonly number[]): Vec3[] {
     .filter((point): point is Vec3 => Boolean(point));
 }
 
-function dot2(a: NormalizedVec2, b: NormalizedVec2): number {
-  return a.x * b.x + a.y * b.y;
-}
-
-function dot3Normalized(a: NormalizedVec3, b: NormalizedVec3): number {
-  return a.x * b.x + a.y * b.y + a.z * b.z;
-}
-
-function cross3Normalized(
-  a: NormalizedVec3,
-  b: NormalizedVec3,
-): NormalizedVec3 {
-  return {
-    x: a.y * b.z - a.z * b.y,
-    y: a.z * b.x - a.x * b.z,
-    z: a.x * b.y - a.y * b.x,
-  };
-}
-
-function pointOnPlacement2D(
-  placement: NormalizedPlacement2D,
-  point: NormalizedVec2,
-): Vec3 {
-  const yAxis = { x: -placement.xAxis.y, y: placement.xAxis.x };
-  return {
-    x:
-      placement.origin.x +
-      point.x * placement.xAxis.x +
-      point.y * yAxis.x,
-    y:
-      placement.origin.y +
-      point.x * placement.xAxis.y +
-      point.y * yAxis.y,
-    z: 0,
-  };
-}
-
-function localPointOnPlacement3D(
-  placement: NormalizedPlacement3D,
-  point: NormalizedVec2,
-): Vec3 {
-  const yAxis = cross3Normalized(placement.zAxis, placement.xAxis);
-  return {
-    x:
-      placement.origin.x +
-      point.x * placement.xAxis.x +
-      point.y * yAxis.x,
-    y:
-      placement.origin.y +
-      point.x * placement.xAxis.y +
-      point.y * yAxis.y,
-    z:
-      placement.origin.z +
-      point.x * placement.xAxis.z +
-      point.y * yAxis.z,
-  };
-}
-
-function pointToPlacement2DLocal(
-  placement: NormalizedPlacement2D,
-  point: Vec3,
-): NormalizedVec2 {
-  const yAxis = { x: -placement.xAxis.y, y: placement.xAxis.x };
-  const relative = {
-    x: point.x - placement.origin.x,
-    y: point.y - placement.origin.y,
-  };
-  return {
-    x: dot2(relative, placement.xAxis),
-    y: dot2(relative, yAxis),
-  };
-}
-
-function pointToPlacement3DLocal(
-  placement: NormalizedPlacement3D,
-  point: Vec3,
-): NormalizedVec2 {
-  const yAxis = cross3Normalized(placement.zAxis, placement.xAxis);
-  const relative = {
-    x: point.x - placement.origin.x,
-    y: point.y - placement.origin.y,
-    z: point.z - placement.origin.z,
-  };
-  return {
-    x: dot3Normalized(relative, placement.xAxis),
-    y: dot3Normalized(relative, yAxis),
-  };
-}
-
-function pointOnCircleOrEllipse(
-  curve: IfcCircle | IfcEllipse,
-  parameter: number,
-): Vec3 {
-  const radiusX = curve.type === "IfcCircle" ? curve.radius : curve.semiAxis1;
-  const radiusY = curve.type === "IfcCircle" ? curve.radius : curve.semiAxis2;
-  const localPoint = {
-    x: radiusX * Math.cos(parameter),
-    y: radiusY * Math.sin(parameter),
-  };
-
-  return curve.position.type === "IfcAxis2Placement2D"
-    ? pointOnPlacement2D(normalizePlacement2D(curve.position), localPoint)
-    : localPointOnPlacement3D(normalizePlacement3D(curve.position), localPoint);
-}
-
-function parameterFromCircleOrEllipsePoint(
-  curve: IfcCircle | IfcEllipse,
-  point: IfcCartesianPoint,
-): number {
-  const worldPoint = cartesianPointToVec3(point);
-  const localPoint =
-    curve.position.type === "IfcAxis2Placement2D"
-      ? pointToPlacement2DLocal(normalizePlacement2D(curve.position), worldPoint)
-      : pointToPlacement3DLocal(normalizePlacement3D(curve.position), worldPoint);
-  const radiusX = curve.type === "IfcCircle" ? curve.radius : curve.semiAxis1;
-  const radiusY = curve.type === "IfcCircle" ? curve.radius : curve.semiAxis2;
-  return Math.atan2(localPoint.y / radiusY, localPoint.x / radiusX);
-}
-
-function sampleCircleOrEllipse(
-  curve: IfcCircle | IfcEllipse,
-  startParameter: number,
-  sweep: number,
-  kind: CurveSegmentKind,
-): ResolvedCurveSegment {
-  const isClosed = Math.abs(Math.abs(sweep) - Math.PI * 2) <= FULL_CIRCLE_EPSILON;
-  const segmentCount = isClosed
-    ? DEFAULT_CONIC_SEGMENTS
-    : Math.max(
-        2,
-        Math.ceil((Math.abs(sweep) / (Math.PI * 2)) * DEFAULT_CONIC_SEGMENTS),
-      );
-  const points: Vec3[] = [];
-
-  for (let index = 0; index <= segmentCount; index += 1) {
-    const ratio = index / segmentCount;
-    addCurvePoint(
-      points,
-      pointOnCircleOrEllipse(curve, startParameter + sweep * ratio),
-      !isClosed,
-    );
-  }
-
-  return { kind, points };
-}
-
-function preferredTrimValue(
-  trim: (IfcCartesianPoint | number)[],
-  masterRepresentation: IfcTrimmedCurve["masterRepresentation"],
-): IfcCartesianPoint | number | undefined {
-  const parameterValue = trim.find((value) => typeof value === "number");
-  const cartesianValue = trim.find((value) => typeof value !== "number");
-
-  if (masterRepresentation === "CARTESIAN") {
-    return cartesianValue ?? parameterValue;
-  }
-  return parameterValue ?? cartesianValue;
-}
-
-function trimValueToParameter(
-  curve: IfcCircle | IfcEllipse,
-  trim: (IfcCartesianPoint | number)[],
-  fallback: number,
-  masterRepresentation: IfcTrimmedCurve["masterRepresentation"],
-): number {
-  const value = preferredTrimValue(trim, masterRepresentation);
-  if (typeof value === "number") {
-    return value;
-  }
-  if (value) {
-    return parameterFromCircleOrEllipsePoint(curve, value);
-  }
-  return fallback;
-}
-
-function resolveCircleOrEllipseSegment(
-  curve: IfcCircle | IfcEllipse,
-  trimmed?: Pick<
-    IfcTrimmedCurve,
-    "trim1" | "trim2" | "senseAgreement" | "masterRepresentation"
-  >,
-): ResolvedCurveSegment {
-  if (!trimmed) {
-    return sampleCircleOrEllipse(curve, 0, Math.PI * 2, "curve");
-  }
-
-  const start = normalizeAngle(
-    trimValueToParameter(
-      curve,
-      trimmed.trim1,
-      0,
-      trimmed.masterRepresentation,
-    ),
-  );
-  const end = normalizeAngle(
-    trimValueToParameter(
-      curve,
-      trimmed.trim2,
-      Math.PI * 2,
-      trimmed.masterRepresentation,
-    ),
-  );
-
-  if (trimmed.senseAgreement) {
-    const adjustedEnd = start >= end ? end + Math.PI * 2 : end;
-    return sampleCircleOrEllipse(curve, start, adjustedEnd - start, "arc");
-  }
-
-  const adjustedStart = start <= end ? start + Math.PI * 2 : start;
-  return sampleCircleOrEllipse(curve, adjustedStart, end - adjustedStart, "arc");
-}
-
 function resolveIndexedPolyCurveSegment(
   controlPoints: Vec3[],
   segment: IfcSegmentIndexSelect,
@@ -441,19 +212,10 @@ export function resolveSupportedCurveSegments(
     case "IfcIndexedPolyCurve":
       return resolveIndexedPolyCurveSegments(curve);
     case "IfcCircle":
-      return [resolveCircleOrEllipseSegment(curve)];
     case "IfcEllipse":
-      return [resolveCircleOrEllipseSegment(curve)];
+      return [resolveConicCurveSegment(curve)];
     case "IfcTrimmedCurve":
-      if (curve.basisCurve.type === "IfcCircle") {
-        return [resolveCircleOrEllipseSegment(curve.basisCurve, curve)];
-      }
-      if (curve.basisCurve.type === "IfcEllipse") {
-        return [resolveCircleOrEllipseSegment(curve.basisCurve, curve)];
-      }
-      throw new Error(
-        `IfcTrimmedCurve basis ${curve.basisCurve.type} is not supported yet`,
-      );
+      return resolveTrimmedCurveSegments(curve);
     case "IfcLine":
       throw new Error(
         "IfcLine is infinite and cannot be rendered without a trimming strategy",
